Separate bower-clean step into different task
Running `bower install` is slow & very prone to failure by error code 128, e.g.

    Fatal error: Failed to execute "git ls-remote --tags --heads git://github.com/angular/bower-angular.git", exit code of #128

Also bower does not have a local cache separate to the project, so removing the
`bower_components` dir means that all bower components have to be redownloaded
from remote repos.

This commit separates bower cleaning into a spearate step, so it can be used
when required but is no longer a default.